Refactor the utils to create the pane to be a class on its own.
